backend/requirements.txt
fastapi
uvicorn
groq
requests
pydantic
python-dotenv
mcp
